+++ src/main/java/pl/smsapi/api/action/hlr/HLRCheckNumber.java
@@ -6,20 +6,27 @@ import pl.smsapi.api.response.CheckNumberResponse;
 
 public class HLRCheckNumber extends AbstractAction<CheckNumberResponse> {
 
+    /**
+     * @deprecated use {@link HLRCheckNumber(String)} instead
+     */
+    @Deprecated
     public HLRCheckNumber() {
         setJson(true);
     }
 
+    public HLRCheckNumber(String phoneNumber) {
+        setJson(true);
+        params.put("number", phoneNumber);
+    }
 
+    /**
+     * @deprecated use {@link HLRCheckNumber(String)} instead
+     */
+    @Deprecated
     public HLRCheckNumber setNumber(String number) {
         params.put("number", number);
         return this;
     }
-
-	/*public HLRCheckNumber setNumber(String[] number) {
-        params.put("number", StringUtils.join(number, ','));
-		return this;
-	}*/
 
     protected CheckNumberResponse createResponse(String data) {
         JSONObject jsonObject = new JSONObject(data);